Update README.md
This is a website designed as per the guidelines of Project0. It is designed on the topic SPACE. Its main page is "Home.html" and all other pages are linked with the help of hyperlinks. I have set an index to reach other pages. Its main stylesheet is "Decor.css". "Style.css" is created as a result of "style.scss"